haruServer/requirements.txt
# FastAPI 核心
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic[email]
python-multipart==0.0.6
sse-starlette==1.8.2

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 认证
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
